+++ packages/orq-rc/src/models/operations/listbudgets.ts
@@ -385,7 +385,7 @@ export const ListBudgetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-12-15T17:21:37.883Z",
+    "2025-12-16T05:40:52.364Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -422,7 +422,7 @@ export const ListBudgetsData$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => ListBudgetsConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-12-15T17:21:37.883Z"))
+  updated: z.date().default(() => new Date("2025-12-16T05:40:52.364Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {