+++ packages/orq-rc/src/models/operations/retrieveconversation.ts
@@ -193,7 +193,7 @@ export const RetrieveConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kchggrrkrk0fh7ytnf231n0q"),
+  _id: z.string().default("conv_01kcjttd2yn3wwdqyd5krvz106"),
   entityId: z.string(),
   kind: RetrieveConversationKind$inboundSchema,
   displayName: z.string(),
@@ -226,7 +226,7 @@ export const RetrieveConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kchggrrkrk0fh7ytnf231n0q"),
+  id: z.string().default("conv_01kcjttd2yn3wwdqyd5krvz106"),
   entityId: z.string(),
   kind: RetrieveConversationKind$outboundSchema,
   displayName: z.string(),