+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -202,7 +202,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KCHGGS5DT235C75K8QKPQW0J"),
+  _id: z.string().default("01KCJTTDGGNXTND21ZXZP1P3XQ"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -245,7 +245,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KCHGGS5DT235C75K8QKPQW0J"),
+  id: z.string().default("01KCJTTDGGNXTND21ZXZP1P3XQ"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,