+++ packages/orq-rc/RELEASES.md
@@ -5589,3 +5589,13 @@ Based on:
 - [typescript v4.1.0-rc.43] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.43 - packages/orq-rc
+
+## 2025-12-16 05:43:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.676.1 (2.781.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.44] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.44 - packages/orq-rc